+++ Tests/CSharp/MemberTests/MemberTests.cs
@@ -297,32 +297,6 @@ internal partial class Class1
         }
 
         [Fact]
-        public async Task TestMethodXmlDocAsync()
-        {
-            await TestConversionVisualBasicToCSharpAsync(
-                @"Class TestClass
-    ''' <summary>Xml doc</summary>
-    Public Sub TestMethod(Of T As {Class, New}, T2 As Structure, T3)(<Out> ByRef argument As T, ByRef argument2 As T2, ByVal argument3 As T3)
-        argument = Nothing
-        argument2 = Nothing
-        argument3 = Nothing
-    End Sub
-End Class", @"
-internal partial class TestClass
-{
-    /// <summary>Xml doc</summary>
-    public void TestMethod<T, T2, T3>(out T argument, ref T2 argument2, T3 argument3)
-        where T : class, new()
-        where T2 : struct
-    {
-        argument = null;
-        argument2 = default;
-        argument3 = default;
-    }
-}");
-        }
-
-        [Fact]
         public async Task TestMethodWithOutParameterAsync()
         {
             await TestConversionVisualBasicToCSharpAsync(